% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendUsagiFileToSTCMtable.R
\name{appendUsagiFileToSTCMtable}
\alias{appendUsagiFileToSTCMtable}
\title{Append Usagi File to Source to Concept Map Table}
\usage{
appendUsagiFileToSTCMtable(
  vocabularyId,
  pathToUsagiFile,
  connection,
  vocabularyDatabaseSchema,
  sourceToConceptMapTable,
  skipValidation = TRUE
)
}
\arguments{
\item{vocabularyId}{String with the vocabulary ID}

\item{pathToUsagiFile}{Path to the Usagi file}

\item{connection}{Database connection object}

\item{vocabularyDatabaseSchema}{Schema name containing vocabulary tables}

\item{sourceToConceptMapTable}{Name of source to concept map table}

\item{skipValidation}{Whether to skip validation of the Usagi file}
}
\description{
Append Usagi File to Source to Concept Map Table
}
