% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoFixDatabaseCodeCountsTable.R
\name{autoFixDatabaseCodeCountsFile}
\alias{autoFixDatabaseCodeCountsFile}
\title{autoFixDatabaseCodeCountTable}
\usage{
autoFixDatabaseCodeCountsFile(
  pathToCodeCountsFile,
  pathToCodeCountsFileFixed = pathToCodeCountsFile,
  keepOnlySourceVocabularyIds = NULL
)
}
\arguments{
\item{pathToCodeCountsFile}{Path to the CSV file containing the database code counts.}

\item{pathToCodeCountsFileFixed}{Path where the fixed CSV file should be saved. Defaults to overwriting the input file.}

\item{keepOnlySourceVocabularyIds}{Vector of source vocabulary IDs to keep. Other source vocabularies will be removed. If NULL, keeps all vocabularies.}
}
\value{
Tibble containing the fixed database code count table.
}
\description{
Automatically fixes issues in the database code count table.
}
\details{
This function performs the following fixes on the provided database code count table:
\enumerate{
\item Removes unused source vocabularies based on the specified \code{keep_only_source_vocabulary_ids}.
\item Sums up repeated values for the same source vocabulary and source code.
\item Removes events with a count less than 5.
}
}
