% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoFixDatabaseCodeCountTable.R
\name{autoFixDatabaseCodeCountTable}
\alias{autoFixDatabaseCodeCountTable}
\title{autoFixDatabaseCodeCountTable}
\usage{
autoFixDatabaseCodeCountTable(
  database_count_table,
  database_name = "",
  keep_only_source_vocabulary_ids = NULL
)
}
\arguments{
\item{keep_only_source_vocabulary_ids}{}
}
\description{
autoFixDatabaseCodeCountTable
}
