% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{buildStatusDashboard}
\alias{buildStatusDashboard}
\title{Build Status Dashboard}
\usage{
buildStatusDashboard(
  mapping_status = NULL,
  output_file_html = file.path(tempdir(), "MappingStatusDashboard.html")
)
}
\arguments{
\item{mapping_status}{Mapping status information.}

\item{output_file_html}{The path to the output HTML file.}
}
\value{
The path to the output HTML file.
}
\description{
This function generates a status dashboard for mapping status.
}
