% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildVocabuliesAll.R
\name{buildVocabulariesAll}
\alias{buildVocabulariesAll}
\title{Run All Validation and Upload Steps}
\usage{
buildVocabulariesAll(
  pathToVocabularyFolder,
  connectionDetails,
  vocabularyDatabaseSchema,
  validationResultsFolder,
  sourceToConceptMapTable = NULL
)
}
\arguments{
\item{pathToVocabularyFolder}{Path to folder containing vocabulary files}

\item{connectionDetails}{DatabaseConnector connection details object}

\item{vocabularyDatabaseSchema}{Schema containing the vocabulary tables}

\item{validationResultsFolder}{Folder where validation results will be saved}

\item{sourceToConceptMapTable}{Optional name of source to concept map table}
}
\description{
Runs the complete workflow of validating vocabulary files and uploading them to CDM tables.
It performs the following steps:
\enumerate{
\item Validate the vocabulary folder
\item If sourceToConceptMapTable is not NULL, create the SourceToConceptMap table
\item Upload the vocabulary.csv and Usagi files to the SourceToConceptMap table
\item Move the SourceToConceptMap table to the CDM table
\item Validate the CDM tables with DQD
}
}
