% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMappingStatus.R
\name{calculateMappingStatus}
\alias{calculateMappingStatus}
\title{calculateMappingStatus}
\usage{
calculateMappingStatus(
  path_to_vocabularies_coverage_file,
  connection_details_omop_tables,
  databases_code_counts_tables,
  ignore_failed_rules = FALSE,
  calculate_all_databases = TRUE
)
}
\arguments{
\item{path_to_vocabularies_coverage_file}{The path to the vocabularies coverage file.}

\item{connection_details_omop_tables}{The connection details for the OMOP tables.}

\item{databases_code_counts_tables}{The tables containing the code counts for different databases.}

\item{ignore_failed_rules}{Logical value indicating whether to ignore databases with failed rules (default is FALSE).}

\item{calculate_all_databases}{Logical value indicating whether to calculate code counts for all databases (default is TRUE).}
}
\value{
A list containing two tables: \code{concepts_to_match} and \code{code_counts_matched}, representing the mapping status for code counts.
}
\description{
Calculates the mapping status for code counts in different databases and vocabularies.
}
\details{
This function calculates the mapping status for code counts in different databases and vocabularies. The calculation involves the following steps:
\enumerate{
\item Validating the input parameters to ensure they meet the required format.
\item Reading and validating the provided vocabularies coverage file.
\item Creating a table called \code{concepts_to_match} that collects concepts in OMOP with mapping relationships and synonyms.
\item Creating a table called \code{all_code_counts} that contains all code counts in the vocabularies coverage, optionally ignoring databases with failed rules.
\item Creating a new database called "all_databases" with normalized event counts across all databases.
\item Calculating the percentage of codes per database and vocabulary.
\item Calculating the mapping status for code counts based on whether they have mappings and codes.
\item Returning the mapping status information in the form of two tables: \code{concepts_to_match} and \code{code_counts_matched}.
}
}
