% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMappingsTablesToOMOPtables.R
\name{convertMappingsTablesToOMOPtables}
\alias{convertMappingsTablesToOMOPtables}
\title{convertMappingsTablesToOMOPtables}
\usage{
convertMappingsTablesToOMOPtables(
  usagi_mapping_tables,
  vocabulary_info_mapping_tables = NULL,
  path_to_temp_omop_vocabulary_folder,
  ignore_failed_rules = FALSE
)
}
\arguments{
\item{ignore_failed_rules}{}
}
\description{
convertMappingsTablesToOMOPtables
}
