% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSourceToConceptMapExtended.R
\name{createSourceToConceptMapExtended}
\alias{createSourceToConceptMapExtended}
\title{Create SOURCE_TO_CONCEPT_MAP_EXTENDED table}
\usage{
createSourceToConceptMapExtended(
  connection,
  vocabularyDatabaseSchema,
  sourceToConceptMapTable,
  overwrite = TRUE
)
}
\arguments{
\item{overwrite}{If TRUE, will drop existing table before creating new one. If FALSE,
will raise an error if table already exists. Default is TRUE.}

\item{connectionDetails}{An R object of type \code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that will contain the OMOP CDM
tables. Requires appropriate user permissions to create new tables.}
}
\value{
TRUE if table is created successfully
}
\description{
Creates the SOURCE_TO_CONCEPT_MAP_EXTENDED table in the specified schema
}
