% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.background_bar_reactable_style}
\alias{.background_bar_reactable_style}
\title{Render a bar chart in the background of the cell}
\usage{
.background_bar_reactable_style(
  string_values,
  colors = NULL,
  height = "60\%",
  width = "95\%",
  background_color = "#FFFFFF",
  text_color = "#FFFFFF"
)
}
\arguments{
\item{string_values}{String with values separated by '-'}

\item{colors}{Vector of colors}

\item{height}{Height of the bar}

\item{width}{Width of the bar}

\item{background_color}{Background color}

\item{text_color}{Text color}
}
\value{
A list of CSS style properties
}
\description{
Render a bar chart in the background of the cell
}
