% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.getCodeCountsForVocabularyAndDatabase}
\alias{.getCodeCountsForVocabularyAndDatabase}
\title{Get code counts for a vocabulary and database}
\usage{
.getCodeCountsForVocabularyAndDatabase(
  pathToCodeCountsFolder,
  sourceVocabularyId,
  databaseName
)
}
\arguments{
\item{pathToCodeCountsFolder}{Path to code counts folder}

\item{sourceVocabularyId}{Source vocabulary ID}

\item{databaseName}{Database name}
}
\value{
Code counts tibble
}
\description{
Get code counts for a vocabulary and database
}
