% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.getDatabaseSummaryForVocabulary}
\alias{.getDatabaseSummaryForVocabulary}
\title{Get database summary for a vocabulary}
\usage{
.getDatabaseSummaryForVocabulary(
  connectionDetails,
  vocabularyDatabaseSchema,
  targetVocabularyIds
)
}
\arguments{
\item{connectionDetails}{Connection details}

\item{vocabularyDatabaseSchema}{Vocabulary database schema}

\item{targetVocabularyIds}{Target vocabulary IDs}
}
\value{
Database summary tibble
}
\description{
Get database summary for a vocabulary
}
