% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.getSummaryListForVocabulary}
\alias{.getSummaryListForVocabulary}
\title{Get summary list for a vocabulary}
\usage{
.getSummaryListForVocabulary(
  pathToCodeCountsFolder,
  pathToVocabularyFolder,
  connectionDetails,
  vocabularyDatabaseSchema,
  sourceVocabularyId,
  targetVocabularyIds
)
}
\arguments{
\item{pathToCodeCountsFolder}{Path to code counts folder}

\item{pathToVocabularyFolder}{Path to vocabulary folder}

\item{connectionDetails}{Connection details}

\item{vocabularyDatabaseSchema}{Vocabulary database schema}

\item{sourceVocabularyId}{Source vocabulary ID}

\item{targetVocabularyIds}{Target vocabulary IDs}
}
\value{
List of summary tables for each database
}
\description{
Get summary list for a vocabulary
}
