% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.getSummaryTableForVocabularyAndDatabase}
\alias{.getSummaryTableForVocabularyAndDatabase}
\title{Get summary table for a vocabulary and database}
\usage{
.getSummaryTableForVocabularyAndDatabase(
  pathToCodeCountsFolder,
  pathToVocabularyFolder,
  connectionDetails,
  vocabularyDatabaseSchema,
  sourceVocabularyId,
  targetVocabularyIds,
  databaseName
)
}
\arguments{
\item{pathToCodeCountsFolder}{Path to code counts folder}

\item{pathToVocabularyFolder}{Path to vocabulary folder}

\item{connectionDetails}{Connection details}

\item{vocabularyDatabaseSchema}{Vocabulary database schema}

\item{sourceVocabularyId}{Source vocabulary ID}

\item{targetVocabularyIds}{Target vocabulary IDs}

\item{databaseName}{Database name}
}
\value{
Summary table as a tibble
}
\description{
Get summary table for a vocabulary and database
}
