% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.getUsagiSummaryForVocabulary}
\alias{.getUsagiSummaryForVocabulary}
\title{Get Usagi summary for a vocabulary}
\usage{
.getUsagiSummaryForVocabulary(pathToVocabularyFolder, sourceVocabularyId)
}
\arguments{
\item{pathToVocabularyFolder}{Path to vocabulary folder}

\item{sourceVocabularyId}{Source vocabulary ID}
}
\value{
Usagi summary tibble
}
\description{
Get Usagi summary for a vocabulary
}
