% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.pageCoverageVocabularyDatabases}
\alias{.pageCoverageVocabularyDatabases}
\title{Render coverage page for a vocabulary and its databases}
\usage{
.pageCoverageVocabularyDatabases(
  summaryTableForVocabularyAndDatabaseList,
  usagiTibble,
  sourceVocabularyId,
  outputFolderPath,
  pathToNewsFile
)
}
\arguments{
\item{summaryTableForVocabularyAndDatabaseList}{List of summary tables}

\item{usagiTibble}{Usagi tibble}

\item{sourceVocabularyId}{Source vocabulary ID}

\item{outputFolderPath}{Output folder path}

\item{pathToNewsFile}{Path to the NEWS.md file}
}
\value{
Path to the output HTML file
}
\description{
Render coverage page for a vocabulary and its databases
}
