% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.pageSummaryAllVocabularies}
\alias{.pageSummaryAllVocabularies}
\title{Render summary page for all vocabularies}
\usage{
.pageSummaryAllVocabularies(summaryAllVocabularies, outputFolderPath)
}
\arguments{
\item{summaryAllVocabularies}{Tibble with summary for all vocabularies}

\item{outputFolderPath}{Output folder path}
}
\value{
Path to the output HTML file
}
\description{
Render summary page for all vocabularies
}
