% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.plotSummaryTableForUsagiFile}
\alias{.plotSummaryTableForUsagiFile}
\title{Plot summary table for Usagi file}
\usage{
.plotSummaryTableForUsagiFile(
  usagiTibble,
  colors = list(flagged = "#EC6173", unchecked = "#F1AE4A", approved = "#51A350", inexact
    = "#AAAAAA")
)
}
\arguments{
\item{usagiTibble}{Usagi tibble}

\item{colors}{List of colors}
}
\value{
A reactable table object
}
\description{
Plot summary table for Usagi file
}
