% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.plotSummaryTableForVocabularyAndDatabase}
\alias{.plotSummaryTableForVocabularyAndDatabase}
\title{Plot summary table for a vocabulary and database}
\usage{
.plotSummaryTableForVocabularyAndDatabase(
  summaryTableForVocabularyAndDatabase,
  colors = list(invalid = "#EC6173", unmapped = "#F1AE4A", mapsTo = "#51A350", grey =
    "#AAAAAA")
)
}
\arguments{
\item{summaryTableForVocabularyAndDatabase}{Tibble with summary}

\item{colors}{List of colors}
}
\value{
A reactable table object
}
\description{
Plot summary table for a vocabulary and database
}
