% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.plotTableForUsagiFile}
\alias{.plotTableForUsagiFile}
\title{Plot Usagi file table}
\usage{
.plotTableForUsagiFile(
  usagiTibble,
  colors = list(flagged = "#EC6173", unchecked = "#F1AE4A", approved = "#51A350", inexact
    = "#AAAAAA")
)
}
\arguments{
\item{usagiTibble}{Usagi tibble}

\item{colors}{List of colors}
}
\value{
A reactable table object
}
\description{
Plot Usagi file table
}
