% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.plotTableForVocabularyAndDatabase}
\alias{.plotTableForVocabularyAndDatabase}
\title{Plot detailed table for a vocabulary and database}
\usage{
.plotTableForVocabularyAndDatabase(
  summaryTableForVocabularyAndDatabase,
  colors = list(invalid = "#EC6173", unmapped = "#F1AE4A", mapsTo = "#51A350", grey =
    "#AAAAAA"),
  fileIssueRepo = ""
)
}
\arguments{
\item{summaryTableForVocabularyAndDatabase}{Tibble with summary}

\item{colors}{List of colors}

\item{fileIssueRepo}{The repository to file issues to}
}
\value{
A reactable table object
}
\description{
Plot detailed table for a vocabulary and database
}
