% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.plotTableSummaryAllVocabularies}
\alias{.plotTableSummaryAllVocabularies}
\title{Plot summary table for all vocabularies}
\usage{
.plotTableSummaryAllVocabularies(
  summaryAllVocabularies,
  colors = list(invalid = "#EC6173", unmapped = "#F1AE4A", mapsTo = "#51A350", grey =
    "#AAAAAA")
)
}
\arguments{
\item{summaryAllVocabularies}{Tibble with summary for all vocabularies}

\item{colors}{List of colors for bar chart}
}
\value{
A reactable table object
}
\description{
Plot summary table for all vocabularies
}
