% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{.tip_reactable_cell}
\alias{.tip_reactable_cell}
\title{Tooltip for reactable cell}
\usage{
.tip_reactable_cell(
  string_values,
  row_names = NULL,
  unit_names = NULL,
  only_percent = FALSE
)
}
\arguments{
\item{string_values}{String with values separated by '-'}

\item{row_names}{Row names for tooltip}

\item{unit_names}{Unit names for tooltip}

\item{only_percent}{Logical, show only percent}
}
\value{
A shiny div with tippy tooltip
}
\description{
Tooltip for reactable cell
}
