% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasesFromAndToCSV.R
\name{duckdbToOMOPVocabularyCSVs}
\alias{duckdbToOMOPVocabularyCSVs}
\title{Export vocabulary tables from a database to CSV files}
\usage{
duckdbToOMOPVocabularyCSVs(
  connection,
  vocabularyDatabaseSchema,
  OMOPVocabularyTableNames,
  pathToOMOPVocabularyCSVsFolder
)
}
\arguments{
\item{connection}{DatabaseConnector connection object to the database containing vocabulary tables}

\item{vocabularyDatabaseSchema}{Schema name where vocabulary tables are located}

\item{OMOPVocabularyTableNames}{Vector of vocabulary table names to export. If NULL, exports standard set.}

\item{pathToOMOPVocabularyCSVsFolder}{Directory path where CSV files will be written}
}
\value{
No return value, called for side effects
}
\description{
Export vocabulary tables from a database to CSV files
}
