% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUpdatedFinOMOPMappings.R
\name{getUpdatedFinOMOPMappings}
\alias{getUpdatedFinOMOPMappings}
\title{get the standard concepts from current FinOMOP vocabulary that are Upgraded/Deprecated in OHDSI latest release}
\usage{
getUpdatedFinOMOPMappings(path_to_omop_vocabulary_folder)
}
\arguments{
\item{path_to_omop_vocabulary_folder}{The path to the folder containing the OMOP vocabulary tables in CSV format.}
}
\value{
list of FinOMOP vocabulary mappings to updated OHDSI latest release
}
\description{
This function creates a temporary database and load the OMOP vocabulary tables from \code{path_to_omop_vocabulary_folder}.
Replace obsolete/discontinued \code{concept_id} using \code{concept_relationship} table
}
