% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper_createATemporaryCopyOfTheOMOPVocabularyDuckDB}
\alias{helper_createATemporaryCopyOfTheOMOPVocabularyDuckDB}
\title{Create a Temporary Copy of the OMOP Vocabulary DuckDB}
\usage{
helper_createATemporaryCopyOfTheOMOPVocabularyDuckDB(
  pathToFullOMOPVocabularyCSVsFolder = NULL
)
}
\arguments{
\item{pathToFullOMOPVocabularyCSVsFolder}{A character string containing the path to the full OMOP vocabulary CSVs folder.}
}
\value{
A character string containing the path to the temporary copy of the DuckDB file
}
\description{
This function creates a temporary copy of the OMOP vocabulary DuckDB file from the package's test data.
}
