% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDatabasesCodeCountsTables.R
\name{importDatabasesCodeCountsTables}
\alias{importDatabasesCodeCountsTables}
\title{importDatabasesCodeCountsTables}
\usage{
importDatabasesCodeCountsTables(
  path_to_database_code_counts_file,
  calculate_all_databases = TRUE
)
}
\arguments{
\item{calculate_all_databases}{}
}
\description{
importDatabasesCodeCountsTables
}
