% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMappingTables.R
\name{importMappingTables}
\alias{importMappingTables}
\title{importMappingTables}
\usage{
importMappingTables(path_to_input_vocabularies_info_file)
}
\arguments{
\item{path_to_input_vocabularies_info_file}{}
}
\description{
importMappingTables
}
