% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeOMOPtables.R
\name{mergeOMOPtables}
\alias{mergeOMOPtables}
\title{mergeOMOPtables}
\usage{
mergeOMOPtables(
  path_to_input_omop_vocabulary_folder,
  path_to_toappend_omop_vocabulary_folder,
  path_to_output_omop_vocabulary_folder
)
}
\arguments{
\item{path_to_input_omop_vocabulary_folder}{path to omop vocabulary tables to merge}

\item{path_to_toappend_omop_vocabulary_folder}{path to omop vocabulary tables to merge}

\item{path_to_output_omop_vocabulary_folder}{path to omop vocabulary tables to output}
}
\description{
merge the omop vocabulary tables in \code{path_to_input_omop_vocabulary_folder} and \code{path_to_input_omop_vocabulary_folder} into a new folder \code{path_to_output_omop_vocabulary_folder}
}
