% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasesFromAndToCSV.R
\name{omopVocabularyCSVsToDuckDB}
\alias{omopVocabularyCSVsToDuckDB}
\title{Convert OMOP Vocabulary CSVs to Database}
\usage{
omopVocabularyCSVsToDuckDB(
  pathToOMOPVocabularyCSVsFolder,
  connection,
  vocabularyDatabaseSchema
)
}
\arguments{
\item{pathToOMOPVocabularyCSVsFolder}{Path to folder containing OMOP vocabulary CSV files}

\item{connection}{A DatabaseConnector connection object}

\item{vocabularyDatabaseSchema}{Schema name where the vocabulary tables are stored}
}
\value{
Nothing
}
\description{
This function converts OMOP vocabulary CSV files to a Database.
}
