% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTableMappingStatus.R
\name{plotTableMappingStatus}
\alias{plotTableMappingStatus}
\title{tableMappingStatus}
\usage{
plotTableMappingStatus(mapping_status)
}
\arguments{
\item{mapping_status}{output from \code{mapOMOPtoCodesFrequencies}}
}
\value{
interactive reactable table
}
\description{
Creates an interactive table from mapping_status data
}
