% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTableVocabularyStatus.R
\name{plotTableVocabularyStatus}
\alias{plotTableVocabularyStatus}
\title{plotTableVocabularyStatus}
\usage{
plotTableVocabularyStatus(mapping_status, source_vocabulary_id)
}
\arguments{
\item{source_vocabulary_id}{}
}
\description{
plotTableVocabularyStatus
}
