% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTable.R
\name{readTable}
\alias{readTable}
\title{readTable}
\usage{
readTable(path_to_table_file, table_type, delim = ",")
}
\arguments{
\item{path_to_table_file}{path to csv file to read}

\item{table_type}{name in \code{tables_valid_format} to read and check the imported table.}

\item{delim}{passed to readr::read_delim, delimiter used in the csv table, Default: ','}
}
\value{
\code{table} read table;
}
\description{
reads a csv table with the column format defined in \code{tables_valid_format}.
}
