% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUsagiFile.R
\name{readUsagiFile}
\alias{readUsagiFile}
\title{Read Usagi File}
\usage{
readUsagiFile(pathToUsagiFile)
}
\arguments{
\item{pathToUsagiFile}{Path to the Usagi mapping file to read}
}
\value{
A tibble containing the Usagi mapping data
}
\description{
Reads a Usagi mapping file with appropriate column type specifications
}
