% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STCMToCDMTables.R
\name{STCMToCDMTables}
\alias{STCMToCDMTables}
\title{Convert Source to Concept Map to CDM Tables}
\usage{
STCMToCDMTables(connection, vocabularyDatabaseSchema, sourceToConceptMapTable)
}
\arguments{
\item{connection}{A DatabaseConnector connection object}

\item{vocabularyDatabaseSchema}{Schema name where the vocabulary tables are stored}

\item{sourceToConceptMapTable}{Name of the source to concept map table}
}
\description{
Converts the contents of a Source to Concept Map table into standard CDM vocabulary tables.
This includes creating entries in:
\itemize{
\item CONCEPT table for source concepts
\item CONCEPT_RELATIONSHIP table for mapping relationships
\item Additional relationship types if using extended STCM format
}
}
