% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateUsagiFile.R
\name{updateUsagiFile}
\alias{updateUsagiFile}
\title{Validate Usagi File}
\usage{
updateUsagiFile(
  pathToUsagiFile,
  connection,
  vocabularyDatabaseSchema,
  pathToUpdatedUsagiFile,
  updateLevelTibble = NULL,
  appendOrClearAutoUpdatingInfo = "append",
  skipValidation = FALSE,
  sourceConceptIdOffset = 0
)
}
\arguments{
\item{pathToUsagiFile}{Path to the Usagi mapping file to validate}

\item{connection}{A DatabaseConnector connection object}

\item{vocabularyDatabaseSchema}{Schema name where the vocabulary tables are stored}

\item{pathToUpdatedUsagiFile}{Path where to save the updated Usagi file}

\item{updateLevelTibble}{Optional tibble defining relationship update rules. Default rules are:
\itemize{
\item "Maps to" (level 1, no review needed)
\item "Concept replaced by" (level 2, review needed)
\item "Concept same_as to" (level 3, review needed)
\item "Concept poss_eq to" (level 4, review needed)
}}

\item{appendOrClearAutoUpdatingInfo}{Whether to append ("append") or clear ("clear") auto-updating info}

\item{skipValidation}{Whether to skip validation checks if TRUE. Default is FALSE}

\item{sourceConceptIdOffset}{Integer offset to add to source concept IDs. Default is 0}
}
\description{
Reads a usagi file given in 'pathToUsagiFile' and performs the following checks:
Default Usagi columns:
\itemize{
\item Check if all default Usagi columns are present:
\item Check if sourceCode and conceptId are unique
\item Check if sourceCode is not empty
\item Check if sourceName is not empty
\item Check if sourceName is less than 255 characters
If usagi file has C&CR columns:
\item Check if concept_id is not 0 for APPROVED mappingStatus
\item Check codes with mapping to more than one domain are mapped to compatible domains
\item Check if sourceValidStartDate is before sourceValidEndDate
\item Check if ADD_INFO:sourceParents is a valid concept code in the ADD_INFO:sourceParentVocabulary
}
}
