% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateCDMtablesWithDQD.R
\name{validateCDMtablesWithDQD}
\alias{validateCDMtablesWithDQD}
\title{Validate OMOP Tables with Data Quality Dashboard}
\usage{
validateCDMtablesWithDQD(
  connectionDetails,
  vocabularyDatabaseSchema,
  validationResultsFolder
)
}
\arguments{
\item{vocabularyDatabaseSchema}{A string with the name of the schema where the vocabulary tables are stored}

\item{connection}{A DatabaseConnector connection}
}
\value{
Results from the Data Quality Dashboard.
}
\description{
This function validates the OMOP vocabulary tables using the Data Quality Dashboard package.
}
