% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateCodeCountsFolder.R
\name{validateCodeCountsFile}
\alias{validateCodeCountsFile}
\title{Validate Code Counts Folder}
\usage{
validateCodeCountsFile(pathToCodeCountsFile)
}
\arguments{
\item{pathToCodeCountsFile}{Path to the CSV file containing the database code counts.}
}
\value{
A tibble containing validation results for all files
}
\description{
Validates all code count files in a code counts folder against a set of rules.
The folder must contain a vocabularies_coverage.csv file that describes which vocabularies
are covered and a databases_coverage.csv file that describes which databases are covered.
}
