% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateOMOPTablesRelationships.R
\name{validateOMOPTablesRelationships}
\alias{validateOMOPTablesRelationships}
\title{validateOMOPTablesRelationships}
\usage{
validateOMOPTablesRelationships(tibble_with_tables)
}
\arguments{
\item{tibble_with_tables}{}
}
\description{
validateOMOPTablesRelationships
}
