% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateOMOPtablesWithDQD.R
\name{validateOMOPtablesWithDQD}
\alias{validateOMOPtablesWithDQD}
\title{Validate OMOP Tables with Data Quality Dashboard}
\usage{
validateOMOPtablesWithDQD(connection_details)
}
\arguments{
\item{connection_details}{A named list of database connection details}
}
\value{
Results from the Data Quality Dashboard.
}
\description{
This function validates the OMOP vocabulary tables using the Data Quality Dashboard package.
}
