% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateTable.R
\name{validateTable}
\alias{validateTable}
\title{validateTable}
\usage{
validateTable(table, table_type)
}
\arguments{
\item{table}{tibble to validate}

\item{table_type}{name in \code{tables_valid_format} to read and check the imported table.}
}
\value{
\code{validation_summary} summary of rules passed;
\code{failed_rules_rows} a tibble with the failed rul name an the corresponding row in \code{table}.
}
\description{
check that a table has the correct columns format.
Vaidation rules are defined in variable \code{tables_valid_format}.
}
