% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversionFile.R
\name{validateUnitConversionTibble}
\alias{validateUnitConversionTibble}
\title{Validate Unit Conversion Tibble}
\usage{
validateUnitConversionTibble(
  unitConversionTibble,
  validUnitsList,
  validQuantitiesList
)
}
\arguments{
\item{unitConversionTibble}{A tibble containing conversion data}

\item{validUnitsList}{A character vector of valid unit codes}

\item{validQuantitiesList}{A character vector of valid quantity codes}
}
\value{
A tibble containing validation results with columns: type, step, message
}
\description{
Validates a conversion tibble against a set of rules:
\itemize{
\item omop_quantity column is not empty and is string
\item source_unit_valid and to_source_unit_valid are not empty, are strings, and all units exist in validUnitsList
\item conversion column is not empty and is string
\item only_to_omop_concepts can be empty and is integer
}
}
