% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateUsagiFile.R
\name{validateUsagiFile}
\alias{validateUsagiFile}
\title{Validate Usagi File}
\usage{
validateUsagiFile(
  pathToUsagiFile,
  connection,
  vocabularyDatabaseSchema,
  pathToValidatedUsagiFile
)
}
\arguments{
\item{pathToUsagiFile}{Path to the Usagi mapping file to validate}

\item{connection}{A DatabaseConnector connection object}

\item{vocabularyDatabaseSchema}{Schema name where the vocabulary tables are stored}

\item{pathToValidatedUsagiFile}{Path where to save the validated Usagi file}
}
\description{
Reads a usagi file given in 'pathToUsagiFile' and performs the following checks:
Default Usagi columns:
\itemize{
\item Check if all default Usagi columns are present:
\item Check if sourceCode and conceptId are unique
\item Check if sourceCode is not empty
\item Check if sourceName is not empty
\item Check if sourceName is less than 255 characters
If usagi file has C&CR columns:
\item Check if concept_id is not 0 for APPROVED mappingStatus
\item Check codes with mapping to more than one domain are mapped to compatible domains
\item Check if sourceValidStartDate is before sourceValidEndDate
\item Check if ADD_INFO:sourceParents is a valid concept code in the ADD_INFO:sourceParentVocabulary
}
}
