% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usagiFile.R
\name{writeUsagiFile}
\alias{writeUsagiFile}
\title{Write a Usagi mapping file}
\usage{
writeUsagiFile(usagiTibble, pathToUsagiFile)
}
\arguments{
\item{usagiTibble}{A tibble containing the Usagi mapping data}

\item{pathToUsagiFile}{Path where the Usagi mapping file should be written}
}
\value{
None
}
\description{
Write a Usagi mapping file to disk
}
\details{
The function writes a Usagi mapping file with the following columns:
\itemize{
\item sourceCode (character): Source code
\item sourceName (character): Source name
\item sourceFrequency (integer): Source frequency
\item sourceAutoAssignedConceptIds (integer): Source auto-assigned concept IDs
\item matchScore (double): Match score
\item mappingStatus (character): Mapping status
\item equivalence (character): Equivalence
\item statusSetBy (character): Status set by
\item statusSetOn (double): Status set on
\item conceptId (integer): Concept ID
\item conceptName (character): Concept name
\item domainId (character): Domain ID
\item mappingType (character): Mapping type
\item comment (character): Comment
\item createdBy (character): Created by
\item createdOn (double): Created on
\item assignedReviewer (character): Assigned reviewer
\item ADD_INFO:sourceConceptId (double): Source concept ID
\item ADD_INFO:sourceConceptClass (character): Source concept class
\item ADD_INFO:sourceDomain (character): Source domain
\item ADD_INFO:sourceValidStartDate (date): Source valid start date
\item ADD_INFO:sourceValidEndDate (date): Source valid end date
\item ADD_INFO:sourceParents (character): Source parents
\item ADD_INFO:sourceParentVocabulary (character): Source parent vocabulary
\item ADD_INFO:validationMessages (character): Validation messages
\item ADD_INFO:autoUpdatingInfo (character): Auto updating info
}
}
