% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{buildStatusDashboard}
\alias{buildStatusDashboard}
\title{Build Status Dashboard}
\usage{
buildStatusDashboard(
  usagi_mapping_tables = NULL,
  vocabulary_info_mapping_tables = NULL,
  results_DQD = NULL,
  databases_code_counts_tables = NULL,
  mapping_status = NULL,
  output_file_html = file.path(tempdir(), "MappingStatusDashboard.html")
)
}
\arguments{
\item{usagi_mapping_tables}{A tibble containing USAGI mapping tables.}

\item{vocabulary_info_mapping_tables}{A tibble containing vocabulary info mapping tables.}

\item{results_DQD}{A list of results from the DQD (Data Quality Dashboard).}

\item{databases_code_counts_tables}{A tibble containing databases code counts tables.}

\item{mapping_status}{Mapping status information.}

\item{output_file_html}{The path to the output HTML file.}
}
\value{
The path to the output HTML file.
}
\description{
This function generates a status dashboard for mapping status.
}
