% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{buildStatusDashboard}
\alias{buildStatusDashboard}
\title{Build Status Dashboard}
\usage{
buildStatusDashboard(
  pathToCodeCountsFolder,
  pathToVocabularyFolder,
  connectionDetails,
  vocabularyDatabaseSchema,
  outputFolderPath = tempdir(),
  fileIssueRepo = ""
)
}
\arguments{
\item{pathToCodeCountsFolder}{Path to folder containing code counts files}

\item{pathToVocabularyFolder}{Path to folder containing vocabulary files}

\item{connectionDetails}{DatabaseConnector connection details object}

\item{vocabularyDatabaseSchema}{Schema containing the vocabulary tables}

\item{outputFolderPath}{The path to the output folder for HTML output}

\item{fileIssueRepo}{The repository to file issues to}
}
\value{
Path to the output HTML file
}
\description{
This function generates a status dashboard for mapping status.
}
