% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{buildStatusDashboard}
\alias{buildStatusDashboard}
\title{buildStatusDashboard}
\usage{
buildStatusDashboard(
  usagi_mapping_tables = NULL,
  vocabulary_info_mapping_tables = NULL,
  results_DQD = NULL,
  databases_code_counts_tables = NULL,
  mapping_status = NULL,
  output_file_html = file.path(tempdir(), "MappingStatusDashboard.html")
)
}
\arguments{
\item{output_file_html}{}
}
\description{
buildStatusDashboard
}
