% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildStatusDashboard.R
\name{buildStatusDashboard}
\alias{buildStatusDashboard}
\title{Build Status Dashboard}
\usage{
buildStatusDashboard(
  pathToCodeCountsFolder,
  connectionDetails,
  vocabularyDatabaseSchema,
  output_file_html = file.path(tempdir(), "MappingStatusDashboard.html")
)
}
\arguments{
\item{pathToCodeCountsFolder}{Path to folder containing code counts files}

\item{connectionDetails}{DatabaseConnector connection details object}

\item{vocabularyDatabaseSchema}{Schema containing the vocabulary tables}

\item{output_file_html}{The path to the output HTML file}
}
\value{
A validation log tibble
}
\description{
This function generates a status dashboard for mapping status.
}
