% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMappingStatus.R
\name{calculateMappingStatus}
\alias{calculateMappingStatus}
\title{calculateMappingStatus}
\usage{
calculateMappingStatus(
  path_to_vocabularies_coverage_file,
  connection_details_omop_tables,
  databases_code_counts_tables,
  ignore_failed_rules = FALSE,
  calculate_all_databases = TRUE
)
}
\arguments{
\item{calculate_all_databases}{}
}
\description{
calculateMappingStatus
}
