% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptRelationshipToAncestorTables.R
\name{conceptRelationshipToAncestorTables}
\alias{conceptRelationshipToAncestorTables}
\title{Convert CONCEPT_RELATIONSHIP to CONCEPT_ANCESTOR}
\usage{
conceptRelationshipToAncestorTables(
  connection,
  vocabularyDatabaseSchema,
  vocabularyList
)
}
\arguments{
\item{connection}{A DatabaseConnector connection object}

\item{vocabularyDatabaseSchema}{Schema containing the vocabulary and STCM tables}

\item{vocabularyList}{Vector of vocabulary_ids to include (default: c("ICD10"))}
}
\value{
Invisible TRUE if successful
}
\description{
Creates and populates the CONCEPT_ANCESTOR table
from CONCEPT_RELATIONSHIP table data.
}
