% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMappingsTablesToOMOPtables.R
\name{convertMappingsTablesToOMOPtables}
\alias{convertMappingsTablesToOMOPtables}
\title{convertMappingsTablesToOMOPtables}
\usage{
convertMappingsTablesToOMOPtables(
  usagi_mapping_tables,
  vocabulary_info_mapping_tables = NULL,
  path_to_temp_omop_vocabulary_folder,
  ignore_failed_rules = FALSE
)
}
\arguments{
\item{usagi_mapping_tables}{Tibble containing the mapping tables imported from Usagi.csv files.}

\item{vocabulary_info_mapping_tables}{Tibble containing the mapping tables imported from info.csv files.}

\item{path_to_temp_omop_vocabulary_folder}{Path to the output folder where to write the OMOP vocabulary tables}

\item{ignore_failed_rules}{Logical value indicating whether to ignore tables with failed rules.}
}
\value{
None
}
\description{
Converts usagi an info mapping tables to OMOP tables.
}
