% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSourceToConceptMapExtended.R
\name{createSourceToConceptMapExtended}
\alias{createSourceToConceptMapExtended}
\title{Create Extended Source to Concept Map Table}
\usage{
createSourceToConceptMapExtended(
  connection,
  vocabularyDatabaseSchema,
  sourceToConceptMapTable,
  overwrite = TRUE
)
}
\arguments{
\item{connection}{A DatabaseConnector connection object}

\item{vocabularyDatabaseSchema}{Schema name where the vocabulary tables are stored}

\item{sourceToConceptMapTable}{Name of the extended source to concept map table to create}

\item{overwrite}{Whether to overwrite an existing table. Default is TRUE}
}
\description{
Creates an extended version of the Source to Concept Map table that includes additional columns
for source concept class, domain, and parent relationships.
}
