% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDatabasesCodeCountsTables.R
\name{importDatabasesCodeCountsTables}
\alias{importDatabasesCodeCountsTables}
\title{Import Databases Code Counts Tables}
\usage{
importDatabasesCodeCountsTables(
  path_to_database_code_counts_file,
  calculate_all_databases = TRUE
)
}
\arguments{
\item{path_to_database_code_counts_file}{The path to the file containing databases code counts.}

\item{calculate_all_databases}{Boolean indicating whether to calculate counts for all databases.}
}
\value{
The imported databases code counts tables.
}
\description{
This function imports individu databases code counts tables.
}
