% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUsagiFile.R
\name{readUsagiFile}
\alias{readUsagiFile}
\title{Read Usagi File}
\usage{
readUsagiFile(pathToUsagiFile)
}
\arguments{
\item{pathToUsagiFile}{Path to the Usagi mapping file to read}
}
\value{
A tibble containing the Usagi mapping data with appropriate column types
}
\description{
Reads a Usagi mapping file with appropriate column type specifications.
Warnings about missing columns are suppressed as the ADD_INFO columns are optional.
}
\details{
Required columns:
\itemize{
\item sourceCode (character): Source vocabulary code
\item sourceName (character): Description of the source code
\item sourceFrequency (integer): Frequency of occurrence
\item sourceAutoAssignedConceptIds (integer): Auto-assigned concept IDs
\item matchScore (double): Score of the mapping match
\item mappingStatus (character): Status of the mapping (APPROVED, UNCHECKED, etc.)
\item equivalence (character): Type of equivalence
\item statusSetBy (character): User who set the status
\item statusSetOn (double): Timestamp when status was set
\item conceptId (integer): Target concept ID
\item conceptName (character): Name of target concept
\item domainId (character): Domain of target concept
\item mappingType (character): Type of mapping
\item comment (character): Additional comments
\item createdBy (character): User who created the mapping
\item createdOn (double): Timestamp of creation
\item assignedReviewer (character): Assigned reviewer
}

Optional ADD_INFO columns:
\itemize{
\item sourceConceptId (double): Source concept ID
\item sourceConceptClass (character): Class of source concept
\item sourceDomain (character): Domain of source concept
\item sourceValidStartDate (date): Start date of validity
\item sourceValidEndDate (date): End date of validity
\item sourceParents (character): Parent concepts
\item sourceParentVocabulary (character): Vocabulary of parent concepts
\item validationMessagess (character): Validation messages
}
}
