% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STCMToCDMTables.R
\name{STCMToCDMTables}
\alias{STCMToCDMTables}
\title{Convert SOURCE_TO_CONCEPT_MAP_EXTENDED to CDM Tables}
\usage{
STCMToCDMTables(connection, vocabularyDatabaseSchema, sourceToConceptMapTable)
}
\arguments{
\item{connection}{A DatabaseConnector connection object}

\item{vocabularyDatabaseSchema}{Schema containing the vocabulary and STCM tables}

\item{sourceToConceptMapTable}{Name of the SOURCE_TO_CONCEPT_MAP_EXTENDED table}
}
\value{
Invisible TRUE if successful
}
\description{
Creates and populates CDM tables (CONCEPT, CONCEPT_RELATIONSHIP, CONCEPT_ANCESTOR)
from SOURCE_TO_CONCEPT_MAP_EXTENDED table data.
}
